backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
web3==6.15.1
python-dotenv==1.0.0
pydantic>=2.7.4
httpx==0.27.0
asyncpg==0.30.0
langchain-core==0.3.51
langchain-openai==0.3.12
openai==1.109.1
